requirements.txt
Flask==1.1.2
pandas==1.0.5
requests==2.24.0
gunicorn
python-dotenv==0.15.0
